Adds a class 'iobase.BoundedSourceSplit' to represent dynamic work rebalancing result of custom sources.
Updates Dataflow runner specific code (apiclient.py) to support dynamic work rebalancing custom sources.

Updates 'OffsetRangeTracker' so that the result of 'position_at_fraction()'' is a 'long' instead of a 'float'.

+++ sdks/python/apache_beam/io/range_trackers.py
@@ -163,7 +163,7 @@ class OffsetRangeTracker(iobase.RangeTracker):
       raise Exception(
           'get_position_for_fraction_consumed is not applicable for an '
           'unbounded range')
-    return (math.ceil(self.start_position() + fraction * (
+    return long(math.ceil(self.start_position() + fraction * (
         self.stop_position() - self.start_position())))
